refactor: Move components and assets into their own folder
