+++ src/assets/img/pause.png
