+++ src/assets/img/search.png
