+++ src/components/Pomodoro.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useState, useEffect, useRef } from 'react';
-import dataService from '../../../services/data';
-import { defaultSettings } from '../../../constants';
-import { Settings } from '../../../types';
+import dataService from '../services/data';
+import { defaultSettings } from '../constants';
+import { Settings } from '../types';
 // add sound effects import audio here
 
 const Pomodoro: React.FC = () => {
