+++ src/pages/Background/index.js
@@ -1,13 +1,18 @@
 console.log('This is the background page.');
 console.log('Put the background scripts here.');
-import { initialData } from '../../services/data';
+import { initialData } from '../../constants';
 
 chrome.sidePanel
   .setPanelBehavior({ openPanelOnActionClick: true })
   .catch((error) => console.error(error));
 
-chrome.storage.sync.set({
-  ...initialData,
+chrome.storage.sync.get('initialized', ({ initialized }) => {
+  if (!initialized) {
+    chrome.storage.sync.set({
+      ...initialData,
+      initialized: true,
+    });
+  }
 });
 
 chrome.storage.onChanged.addListener((changes, namespace) => {
